refactor: Move non-generic things around
