+++ app/charts/column/columns-state.tsx
@@ -18,6 +18,10 @@ import {
 } from "@/charts/column/columns-state-props";
 import { PADDING_INNER, PADDING_OUTER } from "@/charts/column/constants";
 import {
+  ShowColumnValueLabelsVariables,
+  useShowColumnValueLabelsVariables,
+} from "@/charts/column/show-values-utils";
+import {
   getChartWidth,
   useAxisLabelHeightOffset,
   useChartBounds,
@@ -28,8 +32,6 @@ import {
   ChartStateData,
   CommonChartState,
   InteractiveXTimeRangeState,
-  ShowValuesVariables,
-  useShowValuesVariables,
 } from "@/charts/shared/chart-state";
 import { TooltipInfo } from "@/charts/shared/interaction/tooltip";
 import {
@@ -59,7 +61,7 @@ import { ChartProps } from "../shared/ChartProps";
 export type ColumnsState = CommonChartState &
   ColumnsStateVariables &
   InteractiveXTimeRangeState &
-  Omit<ShowValuesVariables, "yOffset"> & {
+  Omit<ShowColumnValueLabelsVariables, "yOffset"> & {
     chartType: "column";
     xScale: ScaleBand<string>;
     xScaleInteraction: ScaleBand<string>;
@@ -237,7 +239,7 @@ const useColumnsState = (
     marginRight: right,
   });
   const { yOffset: yValueLabelsOffset, ...showValuesVariables } =
-    useShowValuesVariables(y, {
+    useShowColumnValueLabelsVariables(y, {
       chartData,
       dimensions: chartProps.dimensions,
       measures: chartProps.measures,
